Task5/requirements.txt
absl-py==2.2.2
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
decorator==4.4.2
easydict==1.13
filelock==3.13.1
fonttools==4.58.0
fsspec==2024.6.1
google-auth==2.40.1
google-auth-oauthlib==1.2.2
grpcio==1.71.0
idna==3.10
imageio==2.31.5
imageio-ffmpeg==0.6.0
Jinja2==3.1.4
kiwisolver==1.4.8
Markdown==3.8
MarkupSafe==2.1.5
matplotlib==3.8.4
moviepy==1.0.3
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pillow==11.0.0
prettytable==3.16.0
proglog==0.1.12
protobuf==4.23.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
scipy==1.12.0
shutup==0.2.0
six==1.17.0
sympy==1.13.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
torch==2.2.2+cu121
torchaudio==2.2.2+cu121
torchvision==0.17.2+cu121
tqdm==4.67.1
triton==2.2.0
typing_extensions==4.12.2
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
